New: Add 'accept' header option on validator
For IIIF services that provide both v2 and v3 manifests at the same URL, the validator needs to send an appropriate 'accept' header to ensure it is returning the appropriate IIIF version response.

This PR adds a checkbox to the validator to optionally include an 'Accept' header in the request. The header is set according to the selected IIIF version, which makes it possible to request v3 manifests from a 'default' v2 service, or vice-versa.